Remove plugin-init.groovy in resources and compile the plugin code with groovyc
